:recycle: refactor: PetInfoView 코드 리팩토링
       body: 중복된 사용으로 인해 따로 PetInfoView 생성
